refactor unit test for service
